secureapprove_django/requirements.txt
# ==================================================
# SecureApprove Django - Requirements
# ==================================================

# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Authentication & Security
django-allauth==0.57.0
django-guardian==2.4.0
python-decouple==3.8
cryptography==41.0.7
# WebAuthn support
webauthn==1.11.1
cbor2==5.4.6

# Database
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0

# File uploads & Static files
Pillow==10.0.1
django-storages==1.14.2

# API & Serialization
django-rest-swagger==2.2.0
drf-yasg==1.21.7
markdown==3.5.1

# Billing & Payments
mercadopago==2.2.1

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Monitoring & Logging
sentry-sdk==1.38.0
django-health-check==3.17.0

# Development
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest-django==4.6.0
polib==1.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
